+++ src/components/Secret.tsx
@@ -42,8 +42,6 @@ export default ({ onSubmit, onDelete, secret }: Props): React.ReactElement => {
       onSubmit={handleSubmit}
       onChange={handleChange}
       formData={data}
-      liveValidate={false}
-      showErrorList={false}
     >
       <Box display='flex' flexDirection='row-reverse' p={1} m={1}>
         <Button variant='contained' color='primary' type='submit' disabled={!dirty} data-cy='button-submit-secret'>