+++ src/layouts/Base.tsx
@@ -172,7 +172,7 @@ export default (props: Props): React.ReactElement => {
       </div>
       <User />
       <Tooltip title='Coming soon: notifications' aria-label='notifications'>
-        <IconButton color='inherit' title='Will soon show notifications!'>
+        <IconButton color='inherit'>
           <Badge badgeContent={1} color='secondary'>
             <NotificationsIcon />
           </Badge>